feat: Add script to insert new blog into MongoDB
This commit adds a new script `mongo.js` that allows inserting a new blog into the MongoDB database. The script takes the password, title, author, url, and likes as command-line arguments and creates a new `Blog` document with the provided values. It then saves the document to the database and prints all existing blogs.

This change is necessary to provide a convenient way to add new blogs to the database for testing and development purposes.